Cast layer tensor as floats

+++ client/aggregation/averaging.py
@@ -12,5 +12,5 @@ class FedAvg(Aggregator):
     def aggregate(self, state_dicts: List[Dict]) -> Dict:
         agg_state = state_dicts[0].copy()
         for key in agg_state:
-            agg_state[key] = torch.stack([state[key] for state in state_dicts if key in state.keys()]).mean(0)
+            agg_state[key] = torch.stack([state[key] for state in state_dicts if key in state.keys()]).float().mean(0)
         return agg_state
